# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaperDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "abstract": {
            "title": "Abstract",
            "type": "string"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "canonical_arxiv_id": {
            "title": "Canonical Arxiv Id",
            "type": "string"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "minimum": 1.0,
            "title": "Current Version",
            "type": "integer"
          },
          "first_submitted_at": {
            "format": "date-time",
            "title": "First Submitted At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_updated_at": {
            "format": "date-time",
            "title": "Latest Updated At",
            "type": "string"
          },
          "pdf_url": {
            "title": "Pdf Url",
            "type": "string"
          },
          "primary_category": {
            "title": "Primary Category",
            "type": "string"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_identities": {
            "items": {
              "$ref": "#/components/schemas/SourceIdentityResponse"
            },
            "title": "Source Identities",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic_slugs": {
            "items": {
              "type": "string"
            },
            "title": "Topic Slugs",
            "type": "array"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/PaperVersionResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "canonical_arxiv_id",
          "title",
          "abstract",
          "current_version",
          "first_submitted_at",
          "latest_updated_at",
          "primary_category",
          "categories",
          "authors",
          "pdf_url",
          "schema_version",
          "created_at",
          "versions",
          "source_identities",
          "topic_slugs"
        ],
        "title": "PaperDetailResponse",
        "type": "object"
      },
      "PaperListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaperListResponse",
        "type": "object"
      },
...
      "PaperSummary": {
        "additionalProperties": false,
        "properties": {
          "abstract": {
            "title": "Abstract",
            "type": "string"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "canonical_arxiv_id": {
            "title": "Canonical Arxiv Id",
            "type": "string"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "minimum": 1.0,
            "title": "Current Version",
            "type": "integer"
          },
          "first_submitted_at": {
            "format": "date-time",
            "title": "First Submitted At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_updated_at": {
            "format": "date-time",
            "title": "Latest Updated At",
            "type": "string"
          },
          "pdf_url": {
            "title": "Pdf Url",
            "type": "string"
          },
          "primary_category": {
            "title": "Primary Category",
            "type": "string"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "canonical_arxiv_id",
          "title",
          "abstract",
          "current_version",
          "first_submitted_at",
          "latest_updated_at",
          "primary_category",
          "categories",
          "authors",
          "pdf_url",
          "schema_version",
          "created_at"
        ],
        "title": "PaperSummary",
        "type": "object"
      },
      "PaperVersionResponse": {
        "additionalProperties": false,
        "properties": {
          "abstract": {
            "title": "Abstract",
            "type": "string"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "canonical_arxiv_id": {
            "title": "Canonical Arxiv Id",
            "type": "string"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "paper_id": {
            "format": "uuid",
            "title": "Paper Id",
            "type": "string"
          },
          "pdf_url": {
            "title": "Pdf Url",
            "type": "string"
          },
          "primary_category": {
            "title": "Primary Category",
            "type": "string"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "paper_id",
          "canonical_arxiv_id",
          "version",
          "title",
          "abstract",
          "submitted_at",
          "updated_at",
          "primary_category",
          "categories",
          "authors",
          "pdf_url",
          "source_url",
          "schema_version",
          "created_at"
        ],
        "title": "PaperVersionResponse",
        "type": "object"
      },
...
      "RunDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor_from": {
            "format": "date-time",
            "title": "Cursor From",
            "type": "string"
          },
          "cursor_to": {
            "format": "date-time",
            "title": "Cursor To",
            "type": "string"
          },
          "discovered_count": {
            "minimum": 0.0,
            "title": "Discovered Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "failed_count": {
            "minimum": 0.0,
            "title": "Failed Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "logical_date": {
            "format": "date",
            "title": "Logical Date",
            "type": "string"
          },
          "normalized_count": {
            "minimum": 0.0,
            "title": "Normalized Count",
            "type": "integer"
          },
          "operation": {
            "$ref": "#/components/schemas/RunOperation"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "topic_id": {
            "format": "uuid",
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic_id",
          "logical_date",
          "operation",
          "status",
          "started_at",
          "completed_at",
          "cursor_from",
          "cursor_to",
          "discovered_count",
          "normalized_count",
          "failed_count",
          "error_code",
          "error_detail",
          "schema_version",
          "created_at",
          "items"
        ],
        "title": "RunDetailResponse",
        "type": "object"
      },
...
      "RunListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunListResponse",
        "type": "object"
      },
      "RunOperation": {
        "enum": [
          "ARXIV_INGESTION"
        ],
        "title": "RunOperation",
        "type": "string"
      },
...
      "RunSummary": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor_from": {
            "format": "date-time",
            "title": "Cursor From",
            "type": "string"
          },
          "cursor_to": {
            "format": "date-time",
            "title": "Cursor To",
            "type": "string"
          },
          "discovered_count": {
            "minimum": 0.0,
            "title": "Discovered Count",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "failed_count": {
            "minimum": 0.0,
            "title": "Failed Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "logical_date": {
            "format": "date",
            "title": "Logical Date",
            "type": "string"
          },
          "normalized_count": {
            "minimum": 0.0,
            "title": "Normalized Count",
            "type": "integer"
          },
          "operation": {
            "$ref": "#/components/schemas/RunOperation"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "topic_id": {
            "format": "uuid",
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic_id",
          "logical_date",
          "operation",
          "status",
          "started_at",
          "completed_at",
          "cursor_from",
          "cursor_to",
          "discovered_count",
          "normalized_count",
          "failed_count",
          "error_code",
          "error_detail",
          "schema_version",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "SourceIdentityResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "paper_id": {
            "format": "uuid",
            "title": "Paper Id",
            "type": "string"
          },
          "paper_version_id": {
            "format": "uuid",
            "title": "Paper Version Id",
            "type": "string"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          },
          "source": {
            "const": "arxiv",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "paper_id",
          "paper_version_id",
          "source",
          "external_id",
          "source_version",
          "source_url",
          "schema_version",
          "created_at"
        ],
        "title": "SourceIdentityResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Private read-oriented API for broad LLM-agent research intelligence.",
    "title": "Domain-Specific Paper Harness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/papers": {
      "get": {
        "operationId": "listPapers",
        "parameters": [
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " List Papers"
      }
    },
    "/api/v1/papers/{paper_id}": {
      "get": {
        "operationId": "getPaper",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Get Paper"
      }
    },
    "/api/v1/runs": {
      "get": {
        "operationId": "listRuns",
        "parameters": [
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " List Runs"
      }
    },
    "/api/v1/runs/latest": {
      "get": {
        "operationId": "getLatestRun",
        "parameters": [
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Latest Run"
      }
    },
...
  }
}